requirements.txt
slackclient
apache-airflow[gcp_api]==1.10.5
